Rebuild ContentSwitch
primary it suppose be shared component for mobile and full screen, but
decided to create two switches, one for mobile, one for fullscreen.Each
one will be nested in different place.There is code redundancy issue,
later on must be extracted and rebuils ad shared component. Not 100%
sure if this approach will work as intended.